pyproject.toml
[project]
name = "code-elephant-ai-tutor"
version = "0.1.0"
description = "AI coding tutor agent with orchestration, RAG, tools, and memory."
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
  "chromadb>=1.0.8",
  "ollama>=0.4.4",
  "pydantic>=2.11.3",
  "python-dotenv>=1.1.0",
  "typer>=0.12.3",
  "rich>=13.9.0",
]

[project.scripts]
tutor = "tutor_agent.main:app"

[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
include = ["tutor_agent*"]
